ai-penetration1/backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
Django==4.2.29
django-axes==8.3.1
django-cors-headers==4.9.0
django-csp==4.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==1.109.1
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-nmap==0.7.1
python-owasp-zap-v2.4==0.1.0
requests==2.33.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zaproxy==0.5.0
